+++ net/message/pong.go
@@ -14,16 +14,16 @@ import (
 
 type pong struct {
 	msgHdr
-	height uint64
+	height uint32
 }
 
 func NewPongMsg() ([]byte, error) {
 	var msg pong
 	msg.msgHdr.Magic = NetID
 	copy(msg.msgHdr.CMD[0:7], "pong")
-	msg.height = uint64(ledger.DefaultLedger.Store.GetHeaderHeight())
+	msg.height = ledger.DefaultLedger.Store.GetHeaderHeight()
 	tmpBuffer := bytes.NewBuffer([]byte{})
-	serialization.WriteUint64(tmpBuffer, msg.height)
+	serialization.WriteUint32(tmpBuffer, msg.height)
 	b := new(bytes.Buffer)
 	err := binary.Write(b, binary.LittleEndian, tmpBuffer.Bytes())
 	if err != nil {
@@ -61,7 +61,7 @@ func (msg pong) Serialize(w io.Writer) error {
 	if err != nil {
 		return err
 	}
-	err = serialization.WriteUint64(w, msg.height)
+	err = serialization.WriteUint32(w, msg.height)
 	if err != nil {
 		return err
 	}
@@ -74,7 +74,7 @@ func (msg *pong) Deserialize(r io.Reader) error {
 	if err != nil {
 		return err
 	}
-	msg.height, err = serialization.ReadUint64(r)
+	msg.height, err = serialization.ReadUint32(r)
 	if err != nil {
 		return err
 	}
